requirements.txt
langchain
langchain-huggingface
langchain-chroma
langchain-groq
sentence-transformers
chromadb
python-dotenv